system-tests: verify machine trusts dCache Test CA
The system test uses certificates issued by the dCache Test CA.  For 
this to work, the client must trust this CA.

Currently there is no verification that the machine has the dCache Test 
CA in the trust store.  If the CA certificates are not trusted then 
transfers will fail.

This patch adds a check that the dCache CA's certificates are trusted.  
If they are not present then the populate script will fail, so failing 
the build.

Target: trunk
Require-notes: no
Require-book: no
Patch: http://rb.dcache.org/r/5122/
Acked-by: Gerd Behrmann