dcap, gridftp: Do not export child cells as well known
DCAP and FTP doors instantiate a child cell per connection. The child
cells are passed the same cell arguments as the parent. Since by default
we export all door cell names as well known (using the -export option),
this means child cells are exported too.

Exporting the child cell names may sometimes be convenient in the admin
shell, however the frequent addition and removal of a well known cell
causes a large number of routing manager updates to be passed between
domains.

In the past, the large number of short lived well known cell names has
been a problem for some broker schemes. Eg the reason wy we do not
simply create a JMS queue per well known cell is that many brokers
treat queues as expensive objects and do not expect thousands to be
created and destroyed every minute.


To the admin the only effect of this change is that the complete cell
address needs to be provided when cd'ing into a child cell of a DCAP
or FTP door.

Target: trunk
Require-book: yes
Require-notes: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/5126/