Missing image resource for poolqueue plots.
It seems that the resource did not get committed with  http://rb.dcache.org/r/5058  (trunk@18016). It needs to be there or else the url for the image used to create a placeholder in the grid view displaying the pool queue plots will be null and Wicket will refuse to render the page.

Target: trunk
Acked-by: Dmitry
Require-notes: no
Require-book: no