axis: better logging of problems
Apache Axis allows for custom dispatch through a class that implements 
the org.apache.axis.Handler interface.  The standard handler, 
RPCProvider, invokes a method on the user-supplied class with the same 
name as the RPC call.  RPCProvider uses a method invokeMethod that 
contains a single call to the method, via reflection.  This allows 
superclasses to decorate how the method call happens; for example, by 
adding additional security checks.

Unfortunately, the error reporting in Axis is less than ideal.  
Exceptions that do not represent SOAP Faults are not logged on the 
server-side, but are reported back to the client as a SOAP Fault.  This 
makes discovering that problems have occurred difficult, let alone 
discovering what caused them and how to fix it.

This class fixes the problem by wrapping the RPC invocation and logging 
activity and problems.  It follows the anti-pattern of catch-log-throw, 
which risks logging the same problem twice; however, this seems to be 
required in order to guarantee that problems are reported in a 
meaningful way.

FOR RELEASE NOTES:

Simply mentioned that dCache SRM now supports better logging should 
there be an unexpected internal error.

Target: trunk
Require-notes: yes
Require-book: no
Patch: http://rb.dcache.org/r/5162/
Acked-by: Gerd Behrmann