webdav: add support for RFC-3230
RFC-3230 describes how a web-server may provide one or more digests (or 
checksums) of a resource as the "Digest" header.  This allows the 
client to determine, when downloading the complete file, whether the 
contents have become corrupted.  RFC-3230 has become important as 
HTTP-based third-party copy needs a mechanism by which the veracity of 
the copy may be verified.  Querying the remote site using the HEAD 
command allows the source to determine the checksum of a recently 
uploaded file.

If the client does not include a 'Want-Digest' then dCache will behave 
as before.  It is only if the client specifies the 'Want-Digest' header 
that dCache fetches the checksum information and provides it as the 
Digest header in the response.

Note that RFC-3230 allows a negotiation between client and server to 
determine which checksum to deliver.  A conforming server may return 
any checksum, including ones that the client did not ask for.  With 
this patch, dCache simply ignores any preference to which algorithm is 
used and responds with all available checksums.

The RFC-3230 representation is returned on all GET and HEAD operations 
on a file.  Returning the RFC-3230 header when delivering partial 
content is probably not that useful, but doing so should not break 
anything and may facilitate some other use-case.  Note that the 
response that redirects the client to the pool does NOT have the 
RFC-3230 header, but the response from the pool will have it.

Unit-tests have been written for the utility method that converts an 
internal dCache representation of a checksum to the corresponding 
RFC-3230 representation.  The existing unit-tests for the pool's HTTP 
response is updated and additional unit tests are added to verify the 
pool provides the correct Digest header.

Also tested using curl and system-test with webdav.redirect.on-read set 
to true and false, and with the client both specifying and omitting the 
'Want-Digest' header.

Note that multi-part read seem broken for curl. For the case when 
relaying through the door, the output is clearly wrong; when 
redirecting, curl hangs.

FOR RELEASE NOTES:

Mention that dCache now makes checksum information available via HTTP 
via RFC-3230.

Target: trunk
Require-notes: yes
Require-book: yes
Patch: http://rb.dcache.org/r/5153
Acked-by: Gerd Behrmann