Add to Alarm key regex group
Currently one can only add the entire message to the alarm key in order to identify it.  This patch allows the regex for alarm definitions to have groups which then can be referenced to be included as part of the key.

An example is given by the modified Checksum alarm (in /etc/dcache/logback.xml):

            <alarmType>
                logger:org.dcache.pool.classic.ChecksumScanner,
                regex:"Checksum mismatch detected for (.+) - marking as BROKEN",
                type:CHECKSUM,
                level:ERROR,
                severity:MODERATE,
                include-in-key:group1 type host service domain
            </alarmType>

Here the tag "group1" in "include-in-key" extracts the pnfsid from the message and includes only that portion of the message string as an identifier.  As usual, group0 is the same as the entire message.

Added ability to match regex against embedded exception messages; added Junit tests for this.

Modified to include "depth" to specify how far down into the embedded Exception messages to match.  Adjusted Junit tests.

Target: trunk
Patch: http://rb.dcache.org/r/5138
Require-book: yes
Require-notes: yes
Acked-by: Dmitry
Request: 2.5

Testing: Reran the modified JUnit tests.  Deployed and tested with ChecksumAlarmSimulator.

RELEASE NOTES:  

Modify the explanation of the alarm key to include matching on regex groups -- see the example above where (.+) represents group 1, and captures the pnfsid in the message.  The tag must be expressed as "group + number" without any whitespace.  Group 0 is identical to "message".  Also added "match-exception" to the filter definition.  False by default, if set to true, applies the regex pattern to all embedded exception messages, recursively, until a match is found.  Added "depth" to be used with "match-exception"; default is undefined (null), meaning unbounded; setting it to an integer > 0 indicates the level to which the match will be applied (in terms of nested messages).  Setting it to 0 is equivalent to setting match-exception to false.