checksum: only accept valid checksums
The Checksum class currently accepts any old rubbish as a value.  The 
internal representation (which should be a hexadecimal value) will 
accept almost any String.  A Checksum object created with an invalid 
value will only becoming apparent later on (when the checksum object is 
used), if at all -- it is possible that an invalid checksum value is 
not noticed.  This violates the "fail fast" principal.

Also, there is an ambiguity in whether the letters in a checksum value 
are represented as upper-case or lower-case.  This results in two 
Checksum objects being unequal if one uses upper-case letters and the 
other lower-case.

This patch adds support for the Checksum class validating and 
normalising the supplied value against the specified ChecksumType.  If 
it isn't valid then IllegalArgumentException is thrown.  The patch 
adjusts the code so it throws NullPointerException if the String 
argument is null.

The code continues to support a special case for ADLER32, where it is 
valid to create a Checksum object with the leading zeros removed.

The normalised checksum value uses lower-case letters as the canonical 
hexadecimal representation.

The unit-test have been substantially improved (one of the two existing 
unit-tests relied on broken behaviour)

Target: trunk
Require-book: no
Require-notes: no
Patch: http://rb.dcache.org/r/5174/
Acked-by: Tigran Mkrtchyan