space-manager: refactor how spaceManagerDisabled is handled
The space-manager has the concept of running while being disabled.  
This allows components that use space-manager as a proxy (for 
pool-manager and the target pool) to be oblivious to whether or not 
space-manager is enabled; when disabled, space-manager still runs and 
forwards some of the messages it receives.

Currently the code for handling the disabled space-manager is spread 
over several methods.  This is unfortunate as the behaviour when 
disabled isn't easy to see and it complicates the methods that operate 
when space-manager is enabled.

This patch consolidates the behaviour to a single method.  This 
simplifies the methods by removing many if-disabled statements that 
distract from the method's purpose.  For the most part, the overall 
behaviour of space-manager is preserved (except that, when 
space-manager is disabled, forwarding of messages happens on the 
message handling thread).  The noticeable expection is that certain 
messages (GetSpaceTokensMessage, GetSpaceTokenIdsMessage, 
GetLinkGroupsMessage, GetLinkGroupIdsMessage) now receive the same 
failure response as other calls to space-manager if space-manager is 
disabled.  Previously, these four methods were processed even when 
space-manager is disabled.

Target: trunk
Require-notes: no
Require-book: no
Patch: http://rb.dcache.org/r/5194/
Acked-by: Dmitry Litvintsev