Make webadmin login more intuitive and friendly
The current webadmin login sequence for admin pages involves two steps which should not be necessary when the user has a valid browser certificate:

    1. actually logging on via the LogIn page
    2. overriding or bypassing the "Guest" login.

This is clumsy and involves unnecessarily repeated operations.  Furthermore, given that gPlazma allows you to map any username to a set of GIDs, and there is nothing that a non-admin user can do on these pages that a Guest user could not, this "Guest" user is actually redundant.

We have thus entirely eliminated (2).  As for (1), we have reinstrumented the redirection so that the following occurs:

1.  If the user has a valid browser cert, and navigates to an admin page, the cert is automatically accessed in case the user is not yet logged in;
2.  If there is no valid browser cert (for Chrome, at least), the cert DN is not mapped, or the mapping provided by the cert does not give the user the role of admin determined by the webadmin GID, then the user is redirected to the login page, where the user may attempt password login (note that failure to login to the admin page through the login page is currently silent: no explicit message such as "permission denied" is given);
3.  If the cert is valid and the user is so mapped, nothing more is needed and the user is allowed to instantiate that page.
4.  The user can at any time log out and log back in using the LogIn page.

Note that this patch does not address the problem of the requirement of a valid cert to establish a secure (https) connection.  It seems that Firefox and Safari will enforce this with our current container setup; Chrome is looser and lets the user through onto the https connection in order to log in.  I will investigate whether it is possible or advisable to generalize the Chrome behavior from the service end, and eventually post a separate patch for that.

Target:  trunk
Patch: http://rb.dcache.org/r/5186
Merge: 2.5
Require-book: no
Require-notes: yes
Acked-by: Dmitry

RELEASE NOTES:  Alert users to the changed behavior of login, whereby a valid cert which is mapped to an admin user allows immediate access to the page without having to go through the login page.

Tested in Firefox, Chrome, Safari (see discussion above).