http: add utility method for decoding RFC 3230 Digest
RFC 3230 describes how to encode and decode checksum values (a.k.a 
digests) in an HTTP response to a GET or HEAD request.  A previous 
patch added support for encoding this information.

This patch adds a utility method for decoding an RFC-3230 Digest 
value to obtain a (possibly empty) set of Checksum objects.  Note 
that this decoder will not throw an exception on badly formatted 
checksum values, but log them at debug level.

Unfortunately the earlier patch used the wrong format for encoding 
the checksums.  This patch fixes that problem.

This patch also removes a utility method (encoding a single Checksum 
object) that is unlikely to be used.

Target: trunk
Require-notes: no
Require-book: no
Patch: http://rb.dcache.org/r/5172/
Acked-by: Tigran Mkrtchyan