chimera: update 'checksum' command to be more useful
The chimera-cli command 'checksum' currently lists all known checksums 
for the targeted file.

This patch adds additional functionality: to list the checksum of a 
specific type, to add a checksum value of a specific type and to remove 
the checksum of a specific type.

The names of checksums are taken from the dCache internal names 
('adler32', 'md5')

Note that there is a problem with the current chimera schema in that 
the ipnfsid field is the primary.  This prevents Chimera from storing 
multiple checksum values against a file; therefore, currently one must 
delete any existing checksum for adding a checksum to work.  Setting 
the primary key to be both ipnfsid and itype will fix this.

Target: trunk
Require-notes: yes
Require-book: yes
Patch: http://rb.dcache.org/r/5222/
Acked-by: Gerd Behrmann