srm: Fix credential delegation for srmCopy transfers
JGlobus 2 changes the GlobusGSSCredentialImpl class and breaks backwards
compatibility of serialization. We include an instance of this class in
RemoteGsiftpTransferProtocolInfo. Thus upgrading to JGlobus 2 will break
compatibility with previous versions.

Since we plan to backport JGlobus 2 to dCache 2.2, this is not just a
concern for pool compatibility, but a concern for head node
compatibility too.

Solving this problem is not easy, as we cannot easily force
serialization of the new GlobusGSSCredentialImpl class to become
compatible with the old format. It stands to reason that relying on the
long term serialization stability of JGlobus classes is a bad idea.

This patch is for 2.5 and 2.6 and must be seen together with patch
http://rb.dcache.org/r/5225/ for 2.2.

Note that dCache with this patch will ONLY be compatible with versions for
which http://rb.dcache.org/r/5225/ has been applied.

Target: trunk
Request: 2.5
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/5227/