SRM: report RuntimeException as bugs
The (abstract) SRM implementation includes the concept of scheduling 
activity.  To achieve this, it has a Scheduler class and various 
subclasses of the abstract Job that represent user-triggered activity.  
Concrete subclasses of Job include a method run, within which is the 
code to enact the user-triggered activity.  The run method may throw 
two checked exceptions: NonFatalJobFailure and FatalJobFailure.  These 
represent a transitory and non-transitory problem, respectively.

There are several problems with the current code:

    many method "throw Exception" rather than being more specific; the 
    classes for handling srmCopy operations are particularly guilty of 
    this.

    RuntimeException (== bugs) are often caught by catch-Exception
    clauses (due to previous problem).  These are then re-thrown as
    NonFatalJobFailure, preventing a stack-trace and letting the job
    retry when it will never succeed.

    If a RuntimeException is thrown by the run method of a subclass
    of Job then the Scheduler class will simply print the error 
    message without a stack-trace.

    In at least one case, a FatalJobFailure is caught and wrapped as a
    NonFatalJobFailure (if user specifies StorageType != DURABLE in 
    srmCopy).

    The run method of some classes (PutFileRequest, 
    ReserveSpaceRequest) update the job state to fail the job 
    manually, rather than throwing FatalJobFailure.

This patch addresses these issues.  It:

    updates methods to declare they throw the specific checked 
    exceptions, rather than Exception.

    updates the catch clauses (mostly in the run methods) to catch 
    specific checked exceptions, allowing RuntimeException and any 
    NonFatalJobFailure and FatalJobFailure exceptions to propagate to 
    Scheduler.

    updates the Scheduler class to catch only NonFatalJobFailure, 
    FatalJobFailure and RuntimeException.  Any RuntimeException 
    results in the job failing and a stack-trace being logged.

    updates the run methods so none attempt to fail the job request; 
    instead, they throw FatalJobFailure and allow Scheduler to fail 
    the job.

Note that SRM treats all errors from the back-end as transitory.  This 
is sub-optimal as the storage system has no mechanism to specify a 
non-transitory failure.  This patch marks problem areas with a 
comment, but doesn't attempt to fix this issue.

Target: trunk
Require-notes: no
Require-book: no
Patch: http://rb.dcache.org/r/5183/
Acked-by: Gerd Behrmann