Alarms Infrastructure Revision (remote end filtering, marking, inclusion of error/warn events) (1 & 2)
This commit combines two patches, the second of which was merely a set of deletes (dead classes) [http://rb.dcache.org/r/5250].

This is the first of two patches (split into two for the sake of readability).  The second patch merely eliminates dead classes (this one must be applied first).

It rectifies objections to current alarm design:

A.  Filtering on sending end (scalability issues);
B.  Inability to raise alarm from inside the code by simply marking the error event;
C.  Current alarm definitions given in JSON (not necessary to break XML syntax);
D.  All error/warn events should be shipped to the remote/alarm server, not just alarms, in order to allow admins to inspect them in one place.

The brunt of the changes, especially in the webadmin module, have to do with renaming and relabeling (since the infrastructure now handles log messages which are not necessarily alarms).  The major changes are as follows:

1.  The logback.xml for the domain now does no filtering per se.  The socket appender sends all events warn and above*** to the alarm server.  It adds three properties to the event MDC before it does.
2.  The storing appender attached to the alarm server now examines the logging events it receives.  It classifies them on the basis of whether there is an ALARM marker and/or a match with one of the custom alarm definitions it knows about.  In the case of an error which is marked as ALARM, the "type" of the alarm will be whatever the marker string is (it should be constructed using the static method AlarmDefinition.getMarker(String subMarker).  All other (non-alarm) entries have a type simply equivalent to their log level (ERROR or WARN).  
3.  The server-side alarm definitions are uniform XML.  JSON appears neither in the configuration nor in the serialization and deserialization of the alarm.
4.  Display of the alarms now allows the user to choose all registered messages, only alarms, and only non-alarms, along with retaining the previous filtering capabilities on severity, type and string search using simple or regex pattern matching (in memory).
5.  There is now a new RDBMs table with an additional attribute.  I will provide a migration script for the rather unlikely case where the RDBMS alarms option has already been used.
6.  The default name for the XML storage has changed from alarms.xml to store.xml.
7.  The server.log is dedicated to ch.qos.logback errors and is non-additive.  All other logging messages go through the storage appender, which then delegates to a history.log (for use with the XML implementation) and an optional SMTP appender.

In terms of the Java code, the following are the most significant changes:

   IAlarmDOA                 --> ILogEntryDAO

   AlarmEntry                --> LogEntry

   AlarmDefinitionAppender  |
                            |--> LogEntryAppender
   AlarmEntryAppender       |

   AlarmDefinition               now is a full setter bean with only default constructor (for use by Joran); a few unnecessary methods have been eliminated and others   
                                 slightly modified.

   The AlarmDefinitionFilter has disappeared.  A new filter, RemoteMDCFilter, which accepts all events but just adds a few properties to the MDC, has been added.

Note also that several places in the code which were accessing a concrete logback object (LoggingEvent) have been changed to access only the interface (ILoggingEvent); this was not only more correct, but particularly necessary in view of the fact that the filtering now is on the receiving end of the socket sender (the socket server), which evidently passes a LoggingEventVO object to the logging API (it is a different implementation of ILoggingEvent from LoggingEvent).

A few adjustments were also made on the webadmin end in order to include the boolean 'alarm' attribute in the queries.

***NOTE: while the current changes are set to send warn-level events to the alarm/remote server, it is my opinion that this makes for a much too verbose log.  Perhaps this will change when we finally rectify the internal logging to conform with the recent suggestions.

Junit tests rewritten (and some eliminated), and run.
Tested on test bed with XML and RDBMS settings.

Target: trunk
Patch:  http://rb.dcache.org/r/5249
Acked-by: Dmitry
Require-notes: yes
Require-book:  yes

RELEASE NOTES and BOOK:  As there a number of significant changes here, I would be willing to help out with the editing.  Feel free to send me emails.