Add JSON API for dCache
The patch provides a response engine for the httpd service. The 
response engine translates users queries to UniversalSpringCell
queries. The query uses reflection to access bean properties. 
It uses a custom serialization scheme to provide a tree of 
public properties (public getters). The response engine 
translates the query response to JSON.

The API is read-only. It is exposed in the default httpd setup
under the api path:

set alias api class org.dcache.services.httpd.probe.ProbeResponseEngine

The API is accessed through HTTP, eg http://localhost:2288/api/PoolManager
or http://localhost:2288/api/PoolManager/psu.activePools.

In contrast to the info service, API requests trigger queries
to the actual service. Thus the information is guaranteed to be
up to date. The downside is that frequent probes would incure
overhead. The API is not guaranteed to be stable at this point (ie
we may change it in between releases). The goal is that any information
that can extracted through the ssh interface is also alvailable through
the API.

Target: trunk
Require-notes: yes
Require-book: yes