wicket: Exclude cglib dependency
Wicket uses cglib for some features in its Spring support.
It relies on cglib 2, which still uses asm 3 that happens
to not be fully compatible with Java 7. We have other
code that relies on asm 4 and we cannot mix the two versions
in dCache.

The patch exludes the cglib dependency. It appears that
webadmin works without it.

Target: trunk
Request: 2.5
Require-book: no
Require-notes: yes
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/5272/