Resolve conflicting imports in Junit test
My Eclipse IDE was complaining about this class.  It seems that there are two 'any' methods, one in org.hamcrest.Matchers and one in org.mockito.Mockito.  With the * import, the reference to any in the class is ambiguous and will be unpredictable.

It is advisable to violate the rule of * for > 3 static imports here, and specify the precise methods used from hamcrest's Matchers class.  Also, since BDDMockito extends Mockito, the additional static imports are not necessary.

Reran test class.

Target: trunk
Patch: http://rb.dcache.org/r/5278
Require-book: no
Require-notes: no
Acked-by: Gerd