ol: Fix migration module regression
The migration module's info command is able to print
the original command used to start the job. The same
functionality is also used to save persistent jobs to
the pool's setup file.

The functionality broken when we refactored the cell
command infrastructure.

This patch solves the problem by relying on a feature
of CellAdapter that allows the currently processed
envelope to be accessed by calling
CellAdapter.getThisMessage(). To be of any use for
cells that do not expose their CellAdapter reference,
I had to change the implementation of getThisMessage()
to use a thread local variable instead.

Target: trunk
Request: 2.5
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/5229/