pool: Upgrade to Berkeley DB version 5.0.58
The new versions is said to contain many bug fixes and performance
improvements.

The file format changed. Pools using the berkeley db backend will be
upgraded to the new format during the first startup. Automatic
downgrade is not possible.

Note that it may be necessary to run a small script to prepare for the
upgrade. If the pool fails to start with an EnvironmentFailureException
failure after upgrading, then the admin needs to run the

   /usr/sbin/dcache-pool-meta-preupgrade

script.

The script and the legacy JE jar file can be removed in dCache 2.7.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/5231/