pool: Fix problem with negative number of active movers
On several pools we recently observed that the number of active movers
reported in the info page and on the httpd pages was negative.

I tracked this down to a standard problem (you may even call it an anti
pattern) with the design of the CompletionHandler callback. The
interface has two completion signals: One for failure and one for
success. What happens if the callback for success fails?

LegacyMoverExecutorService called the failure method when the success
method fails. This leads to the semaphore being released twice, thus
causing a negative number of active jobs.

This patch restructures LegacyMoverExecutorService to no longer call
failed in case completed fails.

An alternative would be to redesign the CompletionHandler to follow the
pattern used by the repository classes with a close/commit/completed
method that is always called no matter whether a failure happened or
not. Then the executor service would look something like this:


try {
  transfer();
} catch (.... e) {
  _completionHandler.failed(e, null);
} finally {
  _completionHandler.completed();
}

Ie completed is no longer a signal of success - it just indicating the
end of the operation. And failed() no longer indicates the end of the
operation - it just indicates an error. This incidentally matches the
implementations of CompletationHandler quite well as the current
implementation of failed() calls completed() as the last statement.


I chose to restructure the callpoint instead to minimize the patch.

Target: trunk
Request: 2.5
Request: 2.2
Request: 1.9.12
Require-book: no
Require-notes: yes
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/5279/