cells: Fix retry loop and logging in sendAndWaitToPermanent
The sendAndWaitToPermanent is a blocking method to send a
message to a cell that is considered permanent. If the cell
is unavailable, this is considered a transient error and
the request is retried until the timeout period has expired.

There is a flaw in the retry logic in that it assumes that
CellNucleus.sendAndWait will return null right away if the
timeout is zero. This is not the case if the
CellGlue.sendMesssage call fails with NoRouteToCellException.
This can lead to an retry loop that never terminates.

There is also a flawed log statement that generates a stack
trace in case of NoRouteToCellException.

Both issues are fixed by this patch.

Target: trunk
Request: 2.5
Request: 2.2
Request: 1.9.12
Require-notes: yes
Require-book: no
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/5284/