Alarm-Definitions (1): Expand AlarmDefinition to work with JDOM Alarm-Definitions (2): Place the Alarm Definitions used by logback into a separate XML file Alarm-Definitions (3a): Add dcache commands for sending alarms, and adding, modifying and removing definitions Alarm-Definitions (3b): Expose available custom alarm definitions using auto-completion text box
(1)
For extensions to the Alarms infrastructure involving the parsing out of the definition file from logback.xml and command-line management of XML definitions, the AlarmsDefinition class has be expanded with several new methods to allow it to be used with JDOM.

Perhaps the best approach would have been to develop a schema for the definition and then use something like JAXB to create Java representations; but that seemed a bit heavy-weight for our purposes.  Hence I built XML representation of the object (which I don't anticipate will be extended) directly into it by including an unmarshaling constructor and a method for marshaling the instance.

There are also validation data structures and methods which will aid in the manipulation of the object via a command-line interpreter.

Some minor changes also were made to the RegexUtils for consistency.

(2)
Currently the server-side custom alarm definitions reside embedded inside the logback-server.xml used by the remote logging service.  This is cumbersome, especially as the list may grow and render the configuration file less readable.  Also, programmatic access to the definitions is more complicated.

This patch changes the LogEntryAppender to read in its definitions from a separate file, now indicated by an alarm property.  The necessary changes to the server wrapper and batch files were made to support this.

There are two other advantages do this separation.  First, it facilitates command-line manipulation of the definitions; second, it facilitates using the predefined definitions to create an auto-complete text box in the query panel of the Alarms.

(3a)
Adds dcache commands for the following:

1.  Send an arbitrary alarm from the commandline to a running alarm server via socket appender;
2.  Add an alarm definition to the alarm-definitions.xml file (via an interpreter);
3.  Remove an alarm definition from the alarm-definitions.xml file (via an interpreter);
4.  Modify an alarm definition in the alarm-definitions.xml file (via an interpreter).

The last three use interactive entry, checking validity immediately after each entry the user makes, displaying the changes before asking for commit, and allowing the user to exit without changing anything.  Errors are not reported as stack traces but as a list of error messages (if there are embedded causes) that are hopefully more readable.

The man pages for dcache contain an explanation of each command.  There are also -h[elp] options for 1, 2, and 3.

(3b)
Makes the Alarm page a little friendlier by reading in the definitions found at alarms.definitions.path and using them to auto-complete the "type" query in the query panel.

Testing done by deployment.

RELEASE NOTES:
BOOK:
The custom, server-side alarm definitions are now found in a separate file, the path of which corresponds to alarms.definitions.path, which defaults (in FHS) to /var/lib/dcache/alarm/alarm-definitions.xml.
It will be sufficient to mention the addition of the dcache commands and what they do.  (For the book, it may or may not be useful to repeat information from the man page.)
Availability of drop-down (auto-complete) choices on web page for alarms. Note that this presupposes, as with the default configuration using the XML event store, that the httpd service has access to the same file system as that used by the alarm server.  It is, of course, always possible to copy the definitions to two different nodes, but there is then the risk that the web admin version may get out of sync.

Target: trunk
Patch: http://rb.dcache.org/r/5302
Patch: http://rb.dcache.org/r/5304
Patch: http://rb.dcache.org/r/5306
Patch: http://rb.dcache.org/r/5307
Acked-by: Tigran
Acked-by: Dmitry
Require-notes: yes
Require-book: yes